requirements.txt
numpy~=1.23.5
pandas~=1.4.2
pathlib~=1.0.1
tqdm~=4.62.3
pillow~=9.4.0
click~=8.1.3
setuptools~=59.5.0
scikit-learn~=1.0.2
imageio~=2.25.1
prettytable~=3.5.0
torchvision~=0.10.1
wandb~=0.13.5